Code Improvements and Changes
Chenged the way platform information is gathered: When a player collides with a platform, we cache the collider of the current platform. If we have not collided with this platform before, we getcomponent and grab the script (rather than store everything in a dictionary).

Also cahnged the leveldata class to store colliders instead of strings. Also separated some more code into smaller helper methods.